added some mostly useless documentation.
Love documentation. Not sure I'd read it, but it feels warm and fuzzy
if there is documentation. Love warm and fuzzy. Anyway, you should
not read this unless you like Doctor Who, latest version. Not
implying this is as funny as the Doctor, just pointing out the state
of mind you need to consume this commit message.