Minimal, not very impressive password generator.
This is better than the brainless welcome thingy. However, this code
still need much more style and functionality. First of all, the length
of the generated password should be configurable via flags, then we need
to support more character sets, and finally we need to configure
different patterns of character sets to use. Oh, and we want to evaluate
the strength of the generated passwords, check them against known words,
and use the proper cryptographic secure random module to add a bit more
pomp to this whole thing.