Implement Screen Full Of Passwords
Following the lead of pwgen, we will generate a screen full of
passwords. We do so by assuming a screen has the historical dimensions
of 80x24 characters.

Note that very long passwords might not be displayed, the screensize is
not really detected, and the number of columns might not scale as
expected.