Add simple command line parser
The simple command line parser, which reads most of pwgens CLI, does
work *except* for the assigning arguments. So you can use "-N 2" to get
2 passwords, but not "--num-passwords=2" (but "--num-passwords 2").
Still work to be done here, but maybe this should be postponed for a
more generic implementation.

The positional arguments are still not evaluated.

Finally, the parser tests are clumsy and incomplete, which is a bad
combination.