Further simplify boolean tests
The boolean CLI tests use reflection to simplify the tests. Now all you
need is specify the flags and the name of the depending variable.
Incidentially, this seems to be a nice way to implement a better CLI
setter logic. Let's see if we can do this in one of the next commits.